+++ backend/src/neuroagent/tools/generate_plot.py
@@ -80,7 +80,7 @@ class PlotGeneratorTool(BaseTool):
     name: ClassVar[str] = "plot-generator"
     name_frontend: ClassVar[str] = "Plot Generator"
     description: ClassVar[str] = (
-        "Generates a plot from user-provided data and saves it to storage. Returns the identifier of the saved plot."
+        "Generates a plot from user-provided data. The plots will be shown after your message automatically."
     )
     description_frontend: ClassVar[
         str