+++ frontend/src/components/plots/barplot.tsx
@@ -11,7 +11,6 @@ import {
   Legend,
 } from "chart.js";
 import { Bar } from "react-chartjs-2";
-
 ChartJS.register(
   CategoryScale,
   LinearScale,
@@ -21,9 +20,11 @@ ChartJS.register(
   Legend,
 );
 import { useGetObjectFromStorage } from "@/hooks/get-storage-object";
-import { urlProp } from "@/lib/types";
+import { PlotProp } from "@/lib/types";
+import Link from "next/link";
+import { Link2 } from "lucide-react";
 
-export function Barplot({ presignedUrl }: urlProp) {
+export function Barplot({ presignedUrl, storageId, isInChat }: PlotProp) {
   const { data: response } = useGetObjectFromStorage(
     presignedUrl as string,
     presignedUrl != "",
@@ -61,10 +62,6 @@ export function Barplot({ presignedUrl }: urlProp) {
       legend: {
         display: false,
       },
-      title: {
-        display: true,
-        text: data.title,
-      },
     },
     scales: {
       y: {
@@ -85,7 +82,14 @@ export function Barplot({ presignedUrl }: urlProp) {
 
   return (
     <div className="w-full max-w-3xl mx-auto p-4 overflow-y-auto">
-      <h2 className="text-xl font-bold mb-2">{data.title}</h2>
+      {isInChat ? (
+        <Link href={`/viewer/${storageId}`} className="flex gap-2">
+          <Link2 className="mt-0.5" />
+          <h2 className="text-xl font-bold mb-2 underline">{data.title}</h2>
+        </Link>
+      ) : (
+        <h2 className="text-xl font-bold mb-2">{data.title}</h2>
+      )}
       {data.description && (
         <p className="text-gray-600 mb-4">{data.description}</p>
       )}