+++ frontend/src/components/plots/piechart.tsx
@@ -9,14 +9,16 @@ import {
   TooltipItem,
 } from "chart.js";
 import { Pie } from "react-chartjs-2";
+import Link from "next/link";
+import { Link2 } from "lucide-react";
 
 // Register the required Chart.js components
 ChartJS.register(ArcElement, Tooltip, Legend);
 
 import { useGetObjectFromStorage } from "@/hooks/get-storage-object";
-import { urlProp } from "@/lib/types";
+import { PlotProp } from "@/lib/types";
 
-export function Piechart({ presignedUrl }: urlProp) {
+export function Piechart({ presignedUrl, isInChat, storageId }: PlotProp) {
   const { data: response } = useGetObjectFromStorage(
     presignedUrl as string,
     presignedUrl != "",
@@ -51,10 +53,7 @@ export function Piechart({ presignedUrl }: urlProp) {
       legend: {
         position: "top" as const,
       },
-      title: {
-        display: true,
-        text: data.title,
-      },
+
       tooltip: {
         callbacks: {
           label: (context: TooltipItem<"pie">) => {
@@ -75,8 +74,15 @@ export function Piechart({ presignedUrl }: urlProp) {
   };
 
   return (
-    <div className="w-full max-w-3xl mx-auto p-4 overflow-y-auto">
-      <h2 className="text-xl font-bold mb-2">{data.title}</h2>
+    <div className={`w-full p-4 overflow-y-auto ${!isInChat && "mx-auto"}`}>
+      {isInChat ? (
+        <Link href={`/viewer/${storageId}`} className="flex gap-2">
+          <Link2 className="mt-0.5" />
+          <h2 className="text-xl font-bold mb-2 underline">{data.title}</h2>
+        </Link>
+      ) : (
+        <h2 className="text-xl font-bold mb-2">{data.title}</h2>
+      )}
       {data.description && (
         <p className="text-gray-600 mb-4">{data.description}</p>
       )}