+++ CHANGELOG.md
@@ -72,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [frontend] - bugfixes.
 - Suggestion endpoint.
 - Support empty vlab+project in threads.
+- [frontend] - tool thunmbnail in chat.
 
 ### Fixed
 - [frontend] - Parallel execution of tools